refactor discovery loop file
